+++ src/Core/Data/Dto/Blog/PostContributionDto.cs
@@ -16,7 +16,7 @@ namespace GamaEdtech.Data.Dto.Blog
         public VisibilityType? VisibilityType { get; set; }
         public DateTimeOffset? PublishDate { get; set; }
         public DateTimeOffset? CreationDate { get; set; }
-        public int? CreationUserId { get; set; }
+        public long? CreationUserId { get; set; }
         public IEnumerable<long>? Tags { get; set; }
         public string? Keywords { get; set; }
         public bool? Draft { get; set; }
