+++ src/Core/Data/Dto/Ticket/ReplyTicketRequestDto.cs
@@ -10,7 +10,7 @@ namespace GamaEdtech.Data.Dto.Ticket
         public required string Body { get; set; }
         public required bool ReplyByAdmin { get; set; }
         public string? From { get; set; }
-        public int? CreationUserId { get; set; }
+        public long? CreationUserId { get; set; }
         public IFormFile? File { get; set; }
         public ICollection<string?>? Receivers { get; set; }
     }
