+++ src/Domain/Entity/Experience.cs
@@ -11,7 +11,7 @@ namespace GamaEdtech.Domain.Entity
     using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
     [Table(nameof(Experience))]
-    public class Experience : IEntity<Experience, long>, IUserId<int>, ISchoolId
+    public class Experience : IEntity<Experience, long>, IUserId<long>, ISchoolId
     {
         [System.ComponentModel.DataAnnotations.Key]
         [Column(nameof(Id), DataType.Long)]
@@ -19,9 +19,9 @@ namespace GamaEdtech.Domain.Entity
         [Required]
         public long Id { get; set; }
 
-        [Column(nameof(UserId), DataType.Int)]
+        [Column(nameof(UserId), DataType.Long)]
         [Required]
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public ApplicationUser User { get; set; }
 
         [Column(nameof(StartDate), DataType.DateTimeOffset)]
